requirements.txt
alembic==1.0.8
blinker==1.4
Bootstrap-Flask==1.0.10
Click==7.0
Flask==1.0.2
Flask-CKEditor==0.4.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.1
Werkzeug==0.14.1
WTForms==2.2.1
